+++ src/main/java/org/chzzk/howmeet/domain/regular/fcm/service/FcmService.java
@@ -12,8 +12,8 @@ import org.chzzk.howmeet.domain.regular.fcm.dto.FcmTokenRequest;
 import org.chzzk.howmeet.domain.regular.fcm.entity.FcmToken;
 import org.chzzk.howmeet.domain.regular.fcm.repository.FcmTokenRepository;
 import org.chzzk.howmeet.domain.regular.notice.entity.Notice;
-import org.chzzk.howmeet.domain.regular.notice.entity.NoticeMessageTemplate;
-import org.chzzk.howmeet.domain.regular.notice.entity.NoticeStatus;
+import org.chzzk.howmeet.domain.regular.notice.model.NoticeMessageTemplate;
+import org.chzzk.howmeet.domain.regular.notice.model.NoticeStatus;
 import org.chzzk.howmeet.domain.regular.notice.repository.NoticeRepository;
 import org.chzzk.howmeet.domain.regular.room.entity.RoomMember;
 import org.chzzk.howmeet.domain.regular.schedule.entity.MemberSchedule;
