+++ src/main/java/org/chzzk/howmeet/domain/regular/record/model/MSRecordNicknames.java
@@ -1,6 +1,5 @@
 package org.chzzk.howmeet.domain.regular.record.model;
 
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
